--- server/src/main/java/seb4141preproject/dto/MultiResponseDto.java
@@ -1,20 +1,0 @@
-package seb4141preproject.dto;
-
-import lombok.Getter;
-import org.springframework.data.domain.Page;
-
-import java.util.List;
-
-@Getter
-public class MultiResponseDto<T> {
-
-    private List<T> data;
-
-    private PageInfo pageInfo;
-
-    public MultiResponseDto(List<T> data, Page page) {
-        this.data = data;
-        this.pageInfo = new PageInfo(page.getNumber() + 1,
-                page.getSize(), page.getTotalElements(), page.getTotalPages());
-    }
-}
