--- server/src/main/java/seb4141preproject/dto/PageInfo.java
@@ -1,13 +1,0 @@
-package seb4141preproject.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class PageInfo {
-    private int page;
-    private int size;
-    private long totalElements;
-    private int totalPages;
-}
