--- server/src/main/java/seb4141preproject/dto/SingleResponseDto.java
@@ -1,10 +1,0 @@
-package seb4141preproject.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class SingleResponseDto<T> {
-    private T data;
-}
